Apple Dividing
Given a number of apples with different weights, how can I split them into two groups such that the difference between the weights of the two groups in minimal? Return this minimum weight.
It took me a day to do this one.
The main problem was integer overflow to to some very heavy apples.