I did one called Creating Strings.
Given a string, your task is to generate all different strings that can be created using its characters.
Input
The only input line has a string of length n. Each character is between a–z.
Output
First print an integer k: the number of strings. Then print k lines: the strings in alphabetical order.

It took me 2 days to do this one.
The main issue was that the recursion was very buggy.