Two Tables
You have an axis-aligned rectangle room with width W
 and height H
, so the lower left corner is in point (0,0)
 and the upper right corner is in (W,H)
.

There is a rectangular table standing in this room. The sides of the table are parallel to the walls, the lower left corner is in (x1,y1)
, and the upper right corner in (x2,y2)
.

You want to place another rectangular table in this room with width w
 and height h
 with the width of the table parallel to the width of the room.

The problem is that sometimes there is not enough space to place the second table without intersecting with the first one (there are no problems with tables touching, though).

You can't rotate any of the tables, but you can move the first table inside the room.

What is the minimum distance you should move the first table to free enough space for the second one?

Input
The first line contains the single integer t
 (1≤t≤5000
) — the number of the test cases.

The first line of each test case contains two integers W
 and H
 (1≤W,H≤108
) — the width and the height of the room.

The second line contains four integers x1
, y1
, x2
 and y2
 (0≤x1<x2≤W
; 0≤y1<y2≤H
) — the coordinates of the corners of the first table.

The third line contains two integers w
 and h
 (1≤w≤W
; 1≤h≤H
) — the width and the height of the second table.

Output
For each test case, print the minimum distance you should move the first table, or −1
 if there is no way to free enough space for the second table.

Your answer will be considered correct if its absolute or relative error doesn't exceed 10−6
.

It took me 1 hour + to do this.
The major problem was that I misread the problem description... :')