Triangles
Farmer John would like to create a triangular pasture for his cows.
There are N
 fence posts (3≤N≤100
) at distinct points (X1,Y1)…(XN,YN)
 on the 2D map of his farm. He can choose three of them to form the vertices of the triangular pasture as long as one of the sides of the triangle is parallel to the x
-axis and another side is parallel to the y
-axis.

What is the maximum area of a pasture that Farmer John can form? It is guaranteed that at least one valid triangular pasture exists.

INPUT FORMAT (file triangles.in):
The first line of the input contains the integer N
. Each of the next N
 lines contains two integers Xi
 and Yi
, each in the range −104…104
 inclusive, describing the location of a fence post.
OUTPUT FORMAT (file triangles.out):
As the area itself is not necessarily an integer, output two times the maximum area of a valid triangle formed by the fence posts.

It took me 1hr 30 to do this.
 I did it on the first try, and there were no problems.