Studying Algorithms
Steph wants to improve her knowledge of algorithms over winter break. She has a total of X
 (1≤X≤104
) minutes to dedicate to learning algorithms. There are N
 (1≤N≤100
) algorithms, and each one of them requires ai
 (1≤ai≤100
) minutes to learn. Find the maximum number of algorithms she can learn.

Input
Line 1: Two space-separated integers N
 and X

Line 2: N
 space-separated integers a1,a2,…aN

Output
A single integer, the answer to the problem.

It took me 30 minutes to do this.
The main problem was that at first I used set, but there are no repeating elements in a set.